producer/requirements.txt
kafka-python==2.2.15
requests-sse